[api/frontend] Enforce RBAC and add a new targets relation (#403, #75)

+++ opencti-platform/opencti-graphql/src/domain/stixDomainEntity.js
@@ -1,4 +1,4 @@
-import { assoc, dissoc, map, propOr, pipe, invertObj } from 'ramda';
+import { assoc, dissoc, map, propOr, pipe, invertObj, isNil } from 'ramda';
 import { BUS_TOPICS } from '../config/conf';
 import { delEditContext, notify, setEditContext } from '../database/redis';
 import {
@@ -23,18 +23,36 @@ import { connectorsForExport } from './connector';
 import { createWork, workToExportFile } from './work';
 import { pushToConnector } from '../database/rabbitmq';
 import stixDomainEntityResolvers from '../resolvers/stixDomainEntity';
+import { ForbiddenAccess } from '../config/errors';
 
-export const findAll = args => {
+export const findAll = async args => {
   const noTypes = !args.types || args.types.length === 0;
   const entityTypes = noTypes ? ['Stix-Domain-Entity'] : args.types;
   const finalArgs = assoc('parentType', 'Stix-Domain-Entity', args);
-  return listEntities(entityTypes, ['name', 'alias'], finalArgs);
+  let data = await listEntities(entityTypes, ['name', 'alias'], finalArgs);
+  data = assoc(
+    'edges',
+    map(
+      n => ({
+        cursor: n.cursor,
+        node: pipe(dissoc('user_email'), dissoc('password'))(n.node),
+        relation: n.relation
+      }),
+      data.edges
+    ),
+    data
+  );
+  return data;
 };
-export const findById = stixDomainEntityId => {
+export const findById = async stixDomainEntityId => {
+  let data = null;
   if (stixDomainEntityId.match(/[a-z-]+--[\w-]{36}/g)) {
-    return loadEntityByStixId(stixDomainEntityId);
+    data = await loadEntityByStixId(stixDomainEntityId, 'Stix-Domain-Entity');
+  } else {
+    data = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
   }
-  return loadEntityById(stixDomainEntityId);
+  data = pipe(dissoc('user_email'), dissoc('password'))(data);
+  return data;
 };
 
 // region time series
@@ -150,7 +168,7 @@ export const stixDomainEntityExportAsk = async args => {
     maxMarkingDefinition = null,
     context = null
   } = args;
-  const entity = stixDomainEntityId ? await loadEntityById(stixDomainEntityId) : null;
+  const entity = stixDomainEntityId ? await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity') : null;
   const workList = await askJobExports(format, entity, type, exportType, maxMarkingDefinition, context, args);
   // Return the work list to do
   return map(w => workToExportFile(w.work), workList);
@@ -177,13 +195,35 @@ export const addStixDomainEntity = async (user, stixDomainEntity) => {
   return notify(BUS_TOPICS.StixDomainEntity.ADDED_TOPIC, created, user);
 };
 export const stixDomainEntityDelete = async stixDomainEntityId => {
-  return deleteEntityById(stixDomainEntityId);
+  const stixDomainEntity = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
+  if (stixDomainEntity.entity_type === 'user' && !isNil(stixDomainEntity.external)) {
+    throw new ForbiddenAccess();
+  }
+  return deleteEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
 };
 export const stixDomainEntityAddRelation = async (user, stixDomainEntityId, input) => {
-  const data = await createRelation(stixDomainEntityId, input);
+  const stixDomainEntity = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
+  if (
+    (stixDomainEntity.entity_type === 'user' &&
+      !isNil(stixDomainEntity.external) &&
+      !['tagged', 'created_by_ref', 'object_marking_refs'].includes(input.through)) ||
+    !input.through
+  ) {
+    throw new ForbiddenAccess();
+  }
+  const data = await createRelation(stixDomainEntityId, input, {}, 'Stix-Domain-Entity', null);
   return notify(BUS_TOPICS.StixDomainEntity.EDIT_TOPIC, data, user);
 };
 export const stixDomainEntityAddRelations = async (user, stixDomainEntityId, input) => {
+  const stixDomainEntity = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
+  if (
+    (stixDomainEntity.entity_type === 'user' &&
+      !isNil(stixDomainEntity.external) &&
+      !['tagged', 'created_by_ref', 'object_marking_refs'].includes(input.through)) ||
+    !input.through
+  ) {
+    throw new ForbiddenAccess();
+  }
   const finalInput = map(
     n => ({
       toId: n,
@@ -193,9 +233,9 @@ export const stixDomainEntityAddRelations = async (user, stixDomainEntityId, inp
     }),
     input.toIds
   );
-  await createRelations(stixDomainEntityId, finalInput);
-  return loadEntityById(stixDomainEntityId).then(stixDomainEntity =>
-    notify(BUS_TOPICS.Workspace.EDIT_TOPIC, stixDomainEntity, user)
+  await createRelations(stixDomainEntityId, finalInput, {}, 'Stix-Domain-Entity', null);
+  return loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity').then(entity =>
+    notify(BUS_TOPICS.Workspace.EDIT_TOPIC, entity, user)
   );
 };
 export const stixDomainEntityDeleteRelation = async (
@@ -203,23 +243,44 @@ export const stixDomainEntityDeleteRelation = async (
   stixDomainEntityId,
   relationId = null,
   toId = null,
-  relationType = 'relation'
+  relationType = 'stix_relation_embedded'
 ) => {
+  const stixDomainEntity = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
   if (relationId) {
+    const data = await loadEntityById(relationId);
+    if (
+      (data.entity_type !== 'stix_relation' && data.entity_type !== 'relation_embedded') ||
+      (stixDomainEntity.entity_type === 'user' &&
+        !isNil(stixDomainEntity.external) &&
+        !['tagged', 'created_by_ref', 'object_marking_refs'].includes(data.relationship_type))
+    ) {
+      throw new ForbiddenAccess();
+    }
     await deleteRelationById(relationId);
   } else if (toId) {
+    if (
+      stixDomainEntity.entity_type === 'user' &&
+      !isNil(stixDomainEntity.external) &&
+      !['tagged', 'created_by_ref', 'object_marking_refs'].includes(relationType)
+    ) {
+      throw new ForbiddenAccess();
+    }
     await deleteRelationsByFromAndTo(stixDomainEntityId, toId, relationType);
   } else {
     throw new Error('Cannot delete the relation, missing relationId or toId');
   }
-  const data = await loadEntityById(stixDomainEntityId);
+  const data = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
   return notify(BUS_TOPICS.StixDomainEntity.EDIT_TOPIC, data, user);
 };
 export const stixDomainEntityEditField = async (user, stixDomainEntityId, input) => {
+  const stixDomainEntity = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
+  if (stixDomainEntity.entity_type === 'user' && !isNil(stixDomainEntity.external)) {
+    throw new ForbiddenAccess();
+  }
   return executeWrite(wTx => {
-    return updateAttribute(stixDomainEntityId, input, wTx);
+    return updateAttribute(stixDomainEntityId, 'Stix-Domain-Entity', input, wTx);
   }).then(async () => {
-    const stixDomain = await loadEntityById(stixDomainEntityId);
+    const stixDomain = await loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity');
     return notify(BUS_TOPICS.StixDomainEntity.EDIT_TOPIC, stixDomain, user);
   });
 };
@@ -228,13 +289,13 @@ export const stixDomainEntityEditField = async (user, stixDomainEntityId, input)
 // region context
 export const stixDomainEntityCleanContext = (user, stixDomainEntityId) => {
   delEditContext(user, stixDomainEntityId);
-  return loadEntityById(stixDomainEntityId).then(stixDomainEntity =>
+  return loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity').then(stixDomainEntity =>
     notify(BUS_TOPICS.StixDomainEntity.EDIT_TOPIC, stixDomainEntity, user)
   );
 };
 export const stixDomainEntityEditContext = (user, stixDomainEntityId, input) => {
   setEditContext(user, stixDomainEntityId, input);
-  return loadEntityById(stixDomainEntityId).then(stixDomainEntity =>
+  return loadEntityById(stixDomainEntityId, 'Stix-Domain-Entity').then(stixDomainEntity =>
     notify(BUS_TOPICS.StixDomainEntity.EDIT_TOPIC, stixDomainEntity, user)
   );
 };
